+++ backend/rs/canisters/tests/src/lib.rs
@@ -1,7 +1,6 @@
-use ic_cdk::update;
-use ic_vetkeys::vetkd_api_types::{
-    VetKDDeriveKeyReply, VetKDDeriveKeyRequest, VetKDKeyId, VetKDPublicKeyReply,
-    VetKDPublicKeyRequest,
+use ic_cdk::{
+    management_canister::{VetKDDeriveKeyArgs, VetKDKeyId, VetKDPublicKeyArgs},
+    update,
 };
 
 #[update]
@@ -25,40 +24,31 @@ async fn vetkd_derive_key(
     key_id: VetKDKeyId,
     transport_public_key: Vec<u8>,
 ) -> Vec<u8> {
-    let request = VetKDDeriveKeyRequest {
+    let request = VetKDDeriveKeyArgs {
         input,
         context,
         key_id,
         transport_public_key,
     };
 
-    let reply: (VetKDDeriveKeyReply,) = ic_cdk::api::call::call_with_payment128(
-        candid::Principal::management_canister(),
-        "vetkd_derive_key",
-        (request,),
-        26_153_846_153,
-    )
-    .await
-    .expect("vetkd_derive_key call failed");
+    let reply = ic_cdk::management_canister::vetkd_derive_key(&request)
+        .await
+        .expect("vetkd_derive_key call failed");
 
-    reply.0.encrypted_key
+    reply.encrypted_key
 }
 
 #[update]
 async fn vetkd_public_key(context: Vec<u8>, key_id: VetKDKeyId) -> Vec<u8> {
-    let request = VetKDPublicKeyRequest {
+    let request = VetKDPublicKeyArgs {
         canister_id: None,
         context,
         key_id,
     };
 
-    let reply: (VetKDPublicKeyReply,) = ic_cdk::api::call::call(
-        candid::Principal::management_canister(),
-        "vetkd_public_key",
-        (request,),
-    )
-    .await
-    .expect("vetkd_public_key call failed");
+    let reply = ic_cdk::management_canister::vetkd_public_key(&request)
+        .await
+        .expect("vetkd_public_key call failed");
 
-    reply.0.public_key
+    reply.public_key
 }