+++ backend/rs/ic_vetkeys/src/lib.rs
@@ -4,7 +4,6 @@
 pub mod encrypted_maps;
 pub mod key_manager;
 pub mod types;
-pub mod vetkd_api_types;
 
 mod utils;
 pub use utils::*;