+++ examples/password_manager_with_metadata/backend/src/lib.rs
@@ -1,4 +1,5 @@
 use candid::{CandidType, Principal};
+use ic_cdk::management_canister::{VetKDCurve, VetKDKeyId};
 use ic_cdk::{init, query, update};
 use ic_stable_structures::memory_manager::{MemoryId, MemoryManager, VirtualMemory};
 use ic_stable_structures::storable::Blob;
@@ -6,7 +7,6 @@ use ic_stable_structures::{storable::Bound, Storable};
 use ic_stable_structures::{BTreeMap as StableBTreeMap, DefaultMemoryImpl};
 use ic_vetkeys::encrypted_maps::{EncryptedMaps, VetKey, VetKeyVerificationKey};
 use ic_vetkeys::types::{AccessRights, ByteBuf, EncryptedMapValue, TransportKey};
-use ic_vetkeys::vetkd_api_types::{VetKDCurve, VetKDKeyId};
 use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
 use std::cell::RefCell;
@@ -101,7 +101,7 @@ fn get_accessible_shared_map_names() -> Vec<(Principal, ByteBuf)> {
         encrypted_maps
             .as_ref()
             .unwrap()
-            .get_accessible_shared_map_names(ic_cdk::caller())
+            .get_accessible_shared_map_names(ic_cdk::api::msg_caller())
             .into_iter()
             .map(|map_id| (map_id.0, ByteBuf::from(map_id.1.as_ref().to_vec())))
             .collect()
@@ -113,7 +113,7 @@ fn get_shared_user_access_for_map(
     map_owner: Principal,
     map_name: ByteBuf,
 ) -> Result<Vec<(Principal, AccessRights)>, String> {
-    let caller = ic_cdk::caller();
+    let caller = ic_cdk::api::msg_caller();
     let key_id = (
         map_owner,
         Blob::try_from(map_name.as_ref()).map_err(|_e| "name too long")?,
@@ -137,7 +137,7 @@ fn get_encrypted_values_for_map_with_metadata(
         encrypted_maps
             .as_ref()
             .unwrap()
-            .get_encrypted_values_for_map(ic_cdk::caller(), map_id)
+            .get_encrypted_values_for_map(ic_cdk::api::msg_caller(), map_id)
     });
     encrypted_values_result.map(|map_values| {
         METADATA.with_borrow(|metadata| {
@@ -167,7 +167,7 @@ fn get_owned_non_empty_map_names() -> Vec<ByteBuf> {
         encrypted_maps
             .as_ref()
             .unwrap()
-            .get_owned_non_empty_map_names(ic_cdk::caller())
+            .get_owned_non_empty_map_names(ic_cdk::api::msg_caller())
             .into_iter()
             .map(|map_name| ByteBuf::from(map_name.as_slice().to_vec()))
             .collect()
@@ -183,7 +183,7 @@ fn insert_encrypted_value_with_metadata(
     tags: Vec<String>,
     url: String,
 ) -> Result<Option<(EncryptedMapValue, PasswordMetadata)>, String> {
-    let caller = ic_cdk::caller();
+    let caller = ic_cdk::api::msg_caller();
     let map_name = bytebuf_to_blob(map_name)?;
     let map_id = (map_owner, map_name);
     let map_key = bytebuf_to_blob(map_key)?;
@@ -218,7 +218,7 @@ fn remove_encrypted_value_with_metadata(
         encrypted_maps
             .as_mut()
             .unwrap()
-            .remove_encrypted_value(ic_cdk::caller(), map_id, map_key)
+            .remove_encrypted_value(ic_cdk::api::msg_caller(), map_id, map_key)
             .map(|opt_prev_value| {
                 METADATA.with_borrow_mut(|metadata| {
                     let metadata_key = (map_owner, map_name, map_key);
@@ -251,7 +251,7 @@ async fn get_encrypted_vetkey(
     Ok(ENCRYPTED_MAPS
         .with_borrow(|encrypted_maps| {
             encrypted_maps.as_ref().unwrap().get_encrypted_vetkey(
-                ic_cdk::caller(),
+                ic_cdk::api::msg_caller(),
                 map_id,
                 transport_key,
             )
@@ -271,7 +271,7 @@ fn get_user_rights(
         encrypted_maps
             .as_ref()
             .unwrap()
-            .get_user_rights(ic_cdk::caller(), map_id, user)
+            .get_user_rights(ic_cdk::api::msg_caller(), map_id, user)
     })
 }
 
@@ -286,7 +286,7 @@ fn set_user_rights(
     let map_id = (map_owner, map_name);
     ENCRYPTED_MAPS.with_borrow_mut(|encrypted_maps| {
         encrypted_maps.as_mut().unwrap().set_user_rights(
-            ic_cdk::caller(),
+            ic_cdk::api::msg_caller(),
             map_id,
             user,
             access_rights,
@@ -306,7 +306,7 @@ fn remove_user(
         encrypted_maps
             .as_mut()
             .unwrap()
-            .remove_user(ic_cdk::caller(), map_id, user)
+            .remove_user(ic_cdk::api::msg_caller(), map_id, user)
     })
 }