+++ backend/rs/ic_vetkeys/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+### Changed
+- Changed error types of `crate::management_canister::{bls_public_key, sign_with_bls}`.
+
 ## [0.1.0] - 27-05-2025
 
 Initial release